+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Report: added album art for the most listened track on the day with most listens.
 - Report: added rating statistics to tracks and albums (average).
 - Report: added loved statistics to tracks and albums (at least 20% loved tracks).
+- Report: added listens by month.
 - Post-Processing: added support for [ghostscript](https://ghostscript.com/) post-processing to optimize the pdf size. Executable (gswin64c.exe or gswin32c.exe) and DLL (gsdll64.dll or gsdll32.dll) must be placed at '.'.\xxx-scripts\helpers-external\ghostscript\'.
 - Post-Processing: added support for configurable CMD commands which are applied to output pdf.
 - Readmes: Ctrl + L. Click on any entry within 'Add button' submenu on toolbar now opens directly their associated readme (without actually adding the button).